refactor: formData separate from formDisplay
